refactor(ui): mv settings components in folder HBL-171
