+++ libs/ui/components/src/lib/Settings/SettingsLogout/index.ts
