+++ libs/ui/components/src/lib/Settings/index.ts
@@ -1,0 +1,5 @@
+import SettingsLogout from './SettingsLogout';
+import SettingsUserInfo from './SettingsUserInfo';
+import { RequiredUserInfoFields } from './SettingsUserInfo';
+
+export { SettingsLogout, SettingsUserInfo, RequiredUserInfoFields };
